Remove Unused Transport and Protocol Files
- Deleted `ITransport.h`, `TransportPseudoCode.h`, and `MCP_Host.h` to streamline the codebase and eliminate unused components.
- This cleanup aligns with ongoing efforts to modernize the transport layer and improve overall maintainability and organization in the codebase.
