+++ Source/Public/Core/Protocol__DT.hpp
@@ -468,290 +468,6 @@ class IMCPProtocol {
             Handler(Error);
         }
     }
-
-  public:
-    // Attaches to the given transport, starts it, and starts listening for messages.
-    // The Protocol object assumes ownership of the Transport, replacing any callbacks that have
-    // already been set, and expects that it is the only user of the Transport instance going
-    // forward.
-    future<void> Connect(const shared_ptr<Transport>& InTransport) {
-        m_Transport = InTransport;
-
-        // Set up transport callbacks
-        m_Transport->OnClose = [this]() { OnCloseInternal(); };
-        m_Transport->OnError = [this](const exception& InError) { OnErrorInternal(InError); };
-        m_Transport->OnMessage = [this](const JSON& InMessage,
-                                        const optional<AuthInfo>& InAuthInfo) {
-            // Parse JSON message and route to appropriate handler
-            if (IsJSONRPCResponse(InMessage) || IsJSONRPCError(InMessage)) {
-                if (IsJSONRPCResponse(InMessage)) {
-                    ResponseBase Response;
-                    Response.JSON_RPC = InMessage[MSG_JSON_RPC];
-                    Response.ID = InMessage[MSG_ID];
-                    Response.Result = InMessage[MSG_RESULT];
-                    OnResponse(response);
-                } else {
-                    ErrorBase ErrorResponse;
-                    ErrorResponse.JSON_RPC = InMessage[MSG_JSON_RPC];
-                    ErrorResponse.ID = InMessage[MSG_ID];
-                    ErrorResponse.Error.Code = InMessage[MSG_ERROR][MSG_CODE];
-                    ErrorResponse.Error.Message = InMessage[MSG_ERROR][MSG_MESSAGE];
-                    if (InMessage[MSG_ERROR].contains(MSG_DATA)) {
-                        ErrorResponse.Error.Data = InMessage[MSG_ERROR][MSG_DATA];
-                    }
-                    OnResponse(error);
-                }
-            } else if (IsJSONRPCRequest(InMessage)) {
-                RequestBase Request;
-                Request.JSON_RPC = InMessage[MSG_JSON_RPC];
-                Request.ID = InMessage[MSG_ID];
-                Request.Method = InMessage[MSG_METHOD];
-                if (InMessage.contains(MSG_PARAMS)) { Request.Params = InMessage[MSG_PARAMS]; }
-                OnRequest(Request, InAuthInfo);
-            } else if (IsJSONRPCNotification(InMessage)) {
-                NotificationBase Notification;
-                Notification.JSON_RPC = InMessage[MSG_JSON_RPC];
-                Notification.Method = InMessage[MSG_METHOD];
-                if (InMessage.contains(MSG_PARAMS)) { Notification.Params = InMessage[MSG_PARAMS]; }
-                OnNotification(Notification);
-            } else {
-                OnErrorInternal(runtime_error("Unknown message type: " + InMessage.dump()));
-            }
-        };
-
-        // Start transport
-        return m_Transport->Start();
-    }
-
-    const shared_ptr<Transport>& GetTransport() const {
-        return m_Transport;
-    }
-
-    // Closes the connection.
-    future<void> Close() {
-        if (m_Transport) { return m_Transport->Close(); }
-        promise<void> Promise;
-        Promise.set_value();
-        return Promise.get_future();
-    }
-
-  protected:
-    // A method to check if a capability is supported by the remote side, for the given method to be
-    // called.
-    // This should be implemented by subclasses.
-    virtual void AssertCapabilityForMethod(const string& InMethod) = 0;
-
-    // A method to check if a notification is supported by the local side, for the given method to
-    // be sent.
-    // This should be implemented by subclasses.
-    virtual void AssertNotificationCapability(const string& InMethod) = 0;
-
-    // A method to check if a request handler is supported by the local side, for the given method
-    // to be handled.
-    // This should be implemented by subclasses.
-    virtual void AssertRequestHandlerCapability(const string& InMethod) = 0;
-
-  public:
-    // Sends a request and wait for a response.
-    // Do not use this method to emit notifications! Use Notification() instead.
-    template <typename ResultT>
-    future<ResultT> Request(const SendRequestT& InRequest,
-                            const optional<RequestOptions>& InOptions = nullopt) {
-        if (!m_Transport) { throw runtime_error("Not connected"); }
-
-        if (InOptions && InOptions->EnforceStrictCapabilities.value_or(false)) {
-            AssertCapabilityForMethod(InRequest.Method);
-        }
-
-        if (InOptions && InOptions->Signal && InOptions->Signal->IsAborted()) {
-            throw runtime_error("Request was aborted");
-        }
-
-        int64_t RequestID = m_RequestRequestID++;
-
-        RequestBase Request;
-        Request.JSON_RPC = MSG_JSON_RPC_VERSION;
-        Request.ID = RequestID;
-        Request.Method = InRequest.Method;
-        Request.Params = InRequest.Params;
-
-        promise<ResultT> ResultPromise;
-        auto ResultFuture = ResultPromise.get_future();
-
-        if (InOptions && InOptions->OnProgress) {
-            lock_guard<mutex> Lock(m_HandlersMutex);
-            m_ProgressHandlers[RequestID] = *InOptions->OnProgress;
-
-            if (!Request.Params) { Request.Params = JSON::object(); }
-            if (!Request.Params->contains(MSG_META)) {
-                (*Request.Params)[MSG_META] = JSON::object();
-            }
-            (*Request.Params)[MSG_META][MSG_PROGRESS_TOKEN] = RequestID;
-        }
-
-        auto Cancel = [this, RequestID](const string& InReason) {
-            {
-                lock_guard<mutex> Lock(m_HandlersMutex);
-                m_ResponseHandlers.erase(RequestID);
-                m_ProgressHandlers.erase(RequestID);
-            }
-            CleanupTimeout(RequestID);
-
-            // Send cancellation notification
-            if (m_Transport) {
-                NotificationBase CancelNotification;
-                CancelNotification.JSON_RPC = MSG_JSON_RPC_VERSION;
-                CancelNotification.Method = MTHD_NOTIFICATION_CANCELLED;
-                CancelNotification.Params = JSON::object();
-                (*CancelNotification.Params)[MSG_REQUEST_ID] = RequestID;
-                (*CancelNotification.Params)["reason"] = InReason;
-
-                m_Transport->Send(CancelNotification);
-            }
-        };
-
-        {
-            lock_guard<mutex> Lock(m_HandlersMutex);
-            m_ResponseHandlers[RequestID] =
-                [ResultPromise = move(ResultPromise), Cancel,
-                 InOptions](const variant<ResponseBase, McpError>& InResponse) mutable {
-                    // TODO: Fix External Ref: AbortSignal - Check if request was aborted
-                    // if (options && options->Signal && options->Signal->IsAborted()) {
-                    //     return;
-                    // }
-
-                    if (holds_alternative<McpError>(InResponse)) {
-                        ResultPromise.set_exception(make_exception_ptr(get<McpError>(InResponse)));
-                        return;
-                    }
-
-                    try {
-                        const auto& Response = get<ResponseBase>(InResponse);
-                        // Parse response result as ResultT using simple JSON conversion
-                        if constexpr (std::is_same_v<ResultT, JSON>) {
-                            ResultPromise.set_value(Response.Result);
-                        } else {
-                            // Simple JSON to type conversion
-                            ResultT Result = Response.Result.get<ResultT>();
-                            ResultPromise.set_value(Result);
-                        }
-                    } catch (const exception& InException) {
-                        ResultPromise.set_exception(current_exception());
-                    }
-                };
-        }
-
-        // TODO: Fix External Ref: AbortSignal - Set up signal cancellation callback
-        // if (options && options->Signal) {
-        //     options->Signal->OnAbort([cancel]() { cancel("Request was cancelled"); });
-        // }
-
-        int64_t Timeout =
-            InOptions && InOptions->Timeout ? *InOptions->Timeout : DEFAULT_REQUEST_TIMEOUT_MSEC;
-        auto TimeoutHandler = [Cancel]() { Cancel("Request timed out"); };
-
-        SetupTimeout(RequestID, Timeout,
-                     InOptions && InOptions->MaxTotalTimeout ? *InOptions->MaxTotalTimeout
-                                                             : nullopt,
-                     TimeoutHandler,
-                     InOptions && InOptions->ResetTimeoutOnProgress.value_or(false)
-                         ? *InOptions->ResetTimeoutOnProgress
-                         : false);
-
-        // Send request via transport
-        // TODO: Relook RequestJSON type
-        try {
-            JSON RequestJSON = {{MSG_JSON_RPC, Request.JSON_RPC},
-                                {MSG_ID, Request.ID},
-                                {MSG_METHOD, Request.Method}};
-            if (Request.Params) { RequestJSON[MSG_PARAMS] = *Request.Params; }
-
-            Transport::TransportSendOptions TransportOptions;
-            if (InOptions) {
-                TransportOptions.RelatedRequestID = InOptions->RelatedRequestID;
-                TransportOptions.ResumptionToken = InOptions->ResumptionToken;
-                TransportOptions.OnResumptionToken = InOptions->OnResumptionToken;
-            }
-
-            m_Transport->Send(RequestJSON, TransportOptions);
-        } catch (const exception& InException) {
-            CleanupTimeout(RequestID);
-            throw;
-        }
-
-        return ResultFuture;
-    }
-
-    // Emits a notification, which is a one-way message that does not expect a response.
-    future<void> Notification(const SendNotificationT& InNotification,
-                              const optional<NotificationOptions>& InOptions = nullopt) {
-        if (!m_Transport) { throw runtime_error("Not connected"); }
-
-        AssertNotificationCapability(InNotification.Method);
-
-        NotificationBase Notification;
-        Notification.JSON_RPC = MSG_JSON_RPC_VERSION;
-        Notification.Method = InNotification.Method;
-        Notification.Params = InNotification.Params;
-
-        JSON NotificationJSON = {{MSG_JSON_RPC, Notification.JSON_RPC},
-                                 {MSG_METHOD, Notification.Method}};
-        if (Notification.Params) { NotificationJSON[MSG_PARAMS] = *Notification.Params; }
-
-        Transport::TransportSendOptions TransportOptions;
-        if (InOptions) { TransportOptions.RelatedRequestID = InOptions->RelatedRequestID; }
-
-        m_Transport->Send(NotificationJSON, TransportOptions);
-
-        promise<void> Promise;
-        Promise.set_value();
-        return Promise.get_future();
-    }
-
-    // Registers a handler to invoke when this protocol object receives a request with the given
-    // method.
-    // Note that this will replace any previous request handler for the same method.
-    void SetRequestHandler(
-        const string& InMethod,
-        function<future<SendResultT>(const RequestBase&,
-                                     const RequestHandlerExtra<SendRequestT, SendNotificationT>&)>
-            InHandler) {
-        AssertRequestHandlerCapability(InMethod);
-
-        lock_guard<mutex> Lock(m_HandlersMutex);
-        m_RequestHandlers[InMethod] = move(InHandler);
-    }
-
-    // Registers a handler to invoke when this protocol object receives a notification with the
-    // given method.
-    // Note that this will replace any previous notification handler for the same method.
-    void SetNotificationHandler(const string& InMethod,
-                                function<future<void>(const NotificationBase&)> InHandler) {
-        lock_guard<mutex> Lock(m_HandlersMutex);
-        m_NotificationHandlers[InMethod] = move(InHandler);
-    }
-
-    // Removes the request handler for the given method.
-    void RemoveRequestHandler(const string& InMethod) {
-        lock_guard<mutex> Lock(m_HandlersMutex);
-        m_RequestHandlers.erase(InMethod);
-    }
-
-    // Asserts that a request handler has not already been set for the given method, in preparation
-    // for a new one being automatically installed.
-    void AssertCanSetRequestHandler(const string& InMethod) {
-        lock_guard<mutex> Lock(m_HandlersMutex);
-        if (m_RequestHandlers.find(InMethod) != m_RequestHandlers.end()) {
-            throw runtime_error("A request handler for " + InMethod
-                                + " already exists, which would be overridden");
-        }
-    }
-
-    // Removes the notification handler for the given method.
-    void RemoveNotificationHandler(const string& InMethod) {
-        lock_guard<mutex> Lock(m_HandlersMutex);
-        m_NotificationHandlers.erase(InMethod);
-    }
 };
 
 template <typename T> T MergeCapabilities(const T& InBase, const T& InAdditional) {
